Deprecation cleanup in flutter_tools (#44469)
One deprecated member was no longer used; removed it.

The other probably shouldn't be deprecated, because it's not the
parser that's deprecated so much as the format it's parsing. The code
itself will live on until we decide to stop supporting the format,
it's not like people calling the member should use something else.