Enable iOS platform views for Flutter Gallery (#45126)
While building the iOS platform views support we only enabled it when the app developer explicitly specified it with a flag, the main reasons were using an untested thread configuration, and taking a performance hit (even when no platform view is used due to disabled raster cache).

Now that we are using the same dynamic thread configuration whether platform views is used or not, and that raster cache works with platform views in the tree as well, we are almost ready to enabled the platform views support by default.

As a first step, I'm enabling it for the Flutter Gallery app on which we run benchmarks. After monitoring the benchmarks and making sure the gallery takes no hit we should be more confident to enabled platform views by default.

This change is expected to have no performance effect on the Gallery app at all.