Enable dev/bots/ build_tests for Linux (#68658)
Enables build_test for Linux. Currently only flutter_gallery has platform directories for the desktop platforms, so this will run only that build, but this will provide an end-to-end build test for Linux.

Other example/test projects can be brought online for Linux in the future just by adding the linux/ directory to the project.