ListTile Material Ripple and Shape Patch (#73618)
This PR replaces the ColoredBox that ListTile uses with an Ink widget. That Ink widget is given a ShapeDecoration with the ListTile's color and shape. This fixes issues where the ListTile color would obscure material ripple effects, and cause the specified shape to not be respected.

List which issues are fixed by this PR. You must list at least one issue.

Fixes #73616
Fixes #63877
Fixes #67117

If you had to change anything in the flutter/tests repo, include a link to the migration guide as per the breaking change policy.

I modified a handful of tests related to ListTiles. The tests I changed had implementation-specific ways of checking the ListTile color. I have rewritten those so that instead of checking for a ColoredBox with a specific color, they check that a path is painted with the correct color.

I added the following tests to list_tile_test.dart:

"ListTile shows Material ripple effects on top of tileColor" (Regression test for #73616)
"ListTile shape is painted correctly" (Regression test for #63877)
I added the following test to sliver_prototype_item_extent_test.dart:

"SliverPrototypeExtentList prototypeItem paint transform is zero" (Regression test for #67117)